+++ grib-writer/src/lib.rs
@@ -10,10 +10,10 @@ use grib_core::binary::{
 };
 use grib_core::bit::BitWriter;
 use grib_core::{
-    ComplexPackingParams, DataRepresentation, FixedSurface, GridDefinition, Identification,
-    ImagePackingParams, Jpeg2000PackingParams, LambertConformalGrid, LatLonGrid, PngPackingParams,
-    PolarStereographicGrid, ProductDefinition, ProductDefinitionTemplate, SimplePackingParams,
-    SpatialDifferencingParams,
+    AlbersEqualAreaGrid, ComplexPackingParams, DataRepresentation, FixedSurface, GridDefinition,
+    Identification, ImagePackingParams, Jpeg2000PackingParams, LambertConformalGrid, LatLonGrid,
+    MercatorGrid, PngPackingParams, PolarStereographicGrid, ProductDefinition,
+    ProductDefinitionTemplate, SimplePackingParams, SpatialDifferencingParams,
 };
 
 pub use grib_core::grib1::ProductDefinition as Grib1ProductDefinition;
@@ -1687,10 +1687,12 @@ fn write_identification_section(out: &mut Vec<u8>, identification: &Identificati
 fn write_grid_section(out: &mut Vec<u8>, grid: &GridDefinition) -> Result<()> {
     match grid {
         GridDefinition::LatLon(grid) => write_latlon_grid_section(out, grid),
+        GridDefinition::Mercator(grid) => write_mercator_grid_section(out, grid),
         GridDefinition::PolarStereographic(grid) => {
             write_polar_stereographic_grid_section(out, grid)
         }
         GridDefinition::LambertConformal(grid) => write_lambert_conformal_grid_section(out, grid),
+        GridDefinition::AlbersEqualArea(grid) => write_albers_equal_area_grid_section(out, grid),
         _ => Err(Error::UnsupportedGridTemplate(grid.template_number())),
     }
 }
@@ -1718,6 +1720,54 @@ fn write_latlon_grid_section(out: &mut Vec<u8>, grid: &LatLonGrid) -> Result<()>
     section[63..67].copy_from_slice(&grid.di.to_be_bytes());
     section[67..71].copy_from_slice(&grid.dj.to_be_bytes());
     section[71] = grid.scanning_mode;
+    out.extend_from_slice(&section);
+    Ok(())
+}
+
+fn write_mercator_grid_section(out: &mut Vec<u8>, grid: &MercatorGrid) -> Result<()> {
+    checked_projected_point_count(grid.ni, grid.nj, grid.number_of_points, "Mercator grid")?;
+
+    let mut section = vec![0u8; 72];
+    section[..4].copy_from_slice(&72u32.to_be_bytes());
+    section[4] = 3;
+    section[6..10].copy_from_slice(&grid.number_of_points.to_be_bytes());
+    section[12..14].copy_from_slice(&10u16.to_be_bytes());
+    write_projected_grid_shape_of_earth(
+        &mut section,
+        ProjectedGridShapeOfEarth {
+            shape_of_earth: grid.shape_of_earth,
+            scale_factor_radius: grid.scale_factor_radius,
+            scaled_value_radius: grid.scaled_value_radius,
+            scale_factor_major_axis: grid.scale_factor_major_axis,
+            scaled_value_major_axis: grid.scaled_value_major_axis,
+            scale_factor_minor_axis: grid.scale_factor_minor_axis,
+            scaled_value_minor_axis: grid.scaled_value_minor_axis,
+        },
+    );
+    section[30..34].copy_from_slice(&grid.ni.to_be_bytes());
+    section[34..38].copy_from_slice(&grid.nj.to_be_bytes());
+    section[38..42].copy_from_slice(&encode_wmo_i32(grid.lat_first).ok_or_else(|| {
+        Error::Other("latitude of first grid point does not fit GRIB signed i32".into())
+    })?);
+    section[42..46].copy_from_slice(&encode_wmo_i32(grid.lon_first).ok_or_else(|| {
+        Error::Other("longitude of first grid point does not fit GRIB signed i32".into())
+    })?);
+    section[46] = grid.resolution_and_component_flags;
+    section[47..51].copy_from_slice(&encode_wmo_i32(grid.lat_d).ok_or_else(|| {
+        Error::Other(
+            "latitude where grid lengths are specified does not fit GRIB signed i32".into(),
+        )
+    })?);
+    section[51..55].copy_from_slice(&encode_wmo_i32(grid.lat_last).ok_or_else(|| {
+        Error::Other("latitude of last grid point does not fit GRIB signed i32".into())
+    })?);
+    section[55..59].copy_from_slice(&encode_wmo_i32(grid.lon_last).ok_or_else(|| {
+        Error::Other("longitude of last grid point does not fit GRIB signed i32".into())
+    })?);
+    section[59] = grid.scanning_mode;
+    section[60..64].copy_from_slice(&grid.orientation_of_grid.to_be_bytes());
+    section[64..68].copy_from_slice(&grid.di.to_be_bytes());
+    section[68..72].copy_from_slice(&grid.dj.to_be_bytes());
     out.extend_from_slice(&section);
     Ok(())
 }
@@ -1767,6 +1817,67 @@ fn write_polar_stereographic_grid_section(
     section[59..63].copy_from_slice(&grid.dy.to_be_bytes());
     section[63] = grid.projection_center_flag;
     section[64] = grid.scanning_mode;
+    out.extend_from_slice(&section);
+    Ok(())
+}
+
+fn write_albers_equal_area_grid_section(
+    out: &mut Vec<u8>,
+    grid: &AlbersEqualAreaGrid,
+) -> Result<()> {
+    checked_projected_point_count(
+        grid.nx,
+        grid.ny,
+        grid.number_of_points,
+        "Albers equal-area grid",
+    )?;
+
+    let mut section = vec![0u8; 81];
+    section[..4].copy_from_slice(&81u32.to_be_bytes());
+    section[4] = 3;
+    section[6..10].copy_from_slice(&grid.number_of_points.to_be_bytes());
+    section[12..14].copy_from_slice(&31u16.to_be_bytes());
+    write_projected_grid_shape_of_earth(
+        &mut section,
+        ProjectedGridShapeOfEarth {
+            shape_of_earth: grid.shape_of_earth,
+            scale_factor_radius: grid.scale_factor_radius,
+            scaled_value_radius: grid.scaled_value_radius,
+            scale_factor_major_axis: grid.scale_factor_major_axis,
+            scaled_value_major_axis: grid.scaled_value_major_axis,
+            scale_factor_minor_axis: grid.scale_factor_minor_axis,
+            scaled_value_minor_axis: grid.scaled_value_minor_axis,
+        },
+    );
+    section[30..34].copy_from_slice(&grid.nx.to_be_bytes());
+    section[34..38].copy_from_slice(&grid.ny.to_be_bytes());
+    section[38..42].copy_from_slice(&encode_wmo_i32(grid.lat_first).ok_or_else(|| {
+        Error::Other("latitude of first grid point does not fit GRIB signed i32".into())
+    })?);
+    section[42..46].copy_from_slice(&grid.lon_first.to_be_bytes());
+    section[46] = grid.resolution_and_component_flags;
+    section[47..51].copy_from_slice(&encode_wmo_i32(grid.lat_d).ok_or_else(|| {
+        Error::Other(
+            "latitude where grid lengths are specified does not fit GRIB signed i32".into(),
+        )
+    })?);
+    section[51..55].copy_from_slice(&grid.lon_v.to_be_bytes());
+    section[55..59].copy_from_slice(&grid.dx.to_be_bytes());
+    section[59..63].copy_from_slice(&grid.dy.to_be_bytes());
+    section[63] = grid.projection_center_flag;
+    section[64] = grid.scanning_mode;
+    section[65..69].copy_from_slice(
+        &encode_wmo_i32(grid.latin1).ok_or_else(|| {
+            Error::Other("first Latin latitude does not fit GRIB signed i32".into())
+        })?,
+    );
+    section[69..73].copy_from_slice(&encode_wmo_i32(grid.latin2).ok_or_else(|| {
+        Error::Other("second Latin latitude does not fit GRIB signed i32".into())
+    })?);
+    section[73..77].copy_from_slice(&encode_wmo_i32(grid.lat_southern_pole).ok_or_else(|| {
+        Error::Other("latitude of southern pole does not fit GRIB signed i32".into())
+    })?);
+    section[77..81].copy_from_slice(&grid.lon_southern_pole.to_be_bytes());
     out.extend_from_slice(&section);
     Ok(())
 }
@@ -2056,6 +2167,9 @@ fn checked_section_length(length: usize, section: u8) -> Result<u32> {
 fn checked_grid_point_count(grid: &GridDefinition) -> Result<usize> {
     match grid {
         GridDefinition::LatLon(grid) => Ok(checked_latlon_point_count(grid)? as usize),
+        GridDefinition::Mercator(grid) => {
+            checked_projected_point_count(grid.ni, grid.nj, grid.number_of_points, "Mercator grid")
+        }
         GridDefinition::PolarStereographic(grid) => checked_projected_point_count(
             grid.nx,
             grid.ny,
@@ -2067,6 +2181,12 @@ fn checked_grid_point_count(grid: &GridDefinition) -> Result<usize> {
             grid.ny,
             grid.number_of_points,
             "Lambert conformal grid",
+        ),
+        GridDefinition::AlbersEqualArea(grid) => checked_projected_point_count(
+            grid.nx,
+            grid.ny,
+            grid.number_of_points,
+            "Albers equal-area grid",
         ),
         _ => Err(Error::UnsupportedGridTemplate(grid.template_number())),
     }
@@ -2144,9 +2264,9 @@ mod tests {
     use grib_core::binary::decode_ibm_f32;
     use grib_core::metadata::ReferenceTime;
     use grib_core::{
-        AnalysisOrForecastTemplate, DataRepresentation, FixedSurface, GridDefinition,
-        Identification, LambertConformalGrid, LatLonGrid, PolarStereographicGrid,
-        ProductDefinition, ProductDefinitionTemplate,
+        AlbersEqualAreaGrid, AnalysisOrForecastTemplate, DataRepresentation, FixedSurface,
+        GridDefinition, Identification, LambertConformalGrid, LatLonGrid, MercatorGrid,
+        PolarStereographicGrid, ProductDefinition, ProductDefinitionTemplate,
     };
     use grib_reader::sections::scan_sections;
     use grib_reader::GribFile;
@@ -2262,6 +2382,31 @@ mod tests {
         })
     }
 
+    fn mercator_grid(scanning_mode: u8) -> GridDefinition {
+        GridDefinition::Mercator(MercatorGrid {
+            number_of_points: 6,
+            shape_of_earth: 6,
+            scale_factor_radius: 0,
+            scaled_value_radius: 0,
+            scale_factor_major_axis: 0,
+            scaled_value_major_axis: 0,
+            scale_factor_minor_axis: 0,
+            scaled_value_minor_axis: 0,
+            ni: 3,
+            nj: 2,
+            lat_first: -20_000_000,
+            lon_first: -100_000_000,
+            resolution_and_component_flags: 0x08,
+            lat_d: 0,
+            lat_last: 20_000_000,
+            lon_last: -98_000_000,
+            scanning_mode,
+            orientation_of_grid: 0,
+            di: 1_000_000,
+            dj: 2_000_000,
+        })
+    }
+
     fn lambert_grid(scanning_mode: u8) -> GridDefinition {
         GridDefinition::LambertConformal(LambertConformalGrid {
             number_of_points: 6,
@@ -2285,6 +2430,34 @@ mod tests {
             scanning_mode,
             latin1: 25_000_000,
             latin2: 25_000_000,
+            lat_southern_pole: -90_000_000,
+            lon_southern_pole: 0,
+        })
+    }
+
+    fn albers_grid(scanning_mode: u8) -> GridDefinition {
+        GridDefinition::AlbersEqualArea(AlbersEqualAreaGrid {
+            number_of_points: 6,
+            shape_of_earth: 6,
+            scale_factor_radius: 0,
+            scaled_value_radius: 0,
+            scale_factor_major_axis: 0,
+            scaled_value_major_axis: 0,
+            scale_factor_minor_axis: 0,
+            scaled_value_minor_axis: 0,
+            nx: 3,
+            ny: 2,
+            lat_first: 23_000_000,
+            lon_first: 240_000_000,
+            resolution_and_component_flags: 0x08,
+            lat_d: 25_000_000,
+            lon_v: 265_000_000,
+            dx: 4_000_000,
+            dy: 5_000_000,
+            projection_center_flag: 0,
+            scanning_mode,
+            latin1: 29_500_000,
+            latin2: 45_500_000,
             lat_southern_pole: -90_000_000,
             lon_southern_pole: 0,
         })
@@ -2526,6 +2699,48 @@ mod tests {
     }
 
     #[test]
+    fn writes_mercator_grib2_field_readable_by_reader() {
+        let values = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0];
+        let field = Grib2FieldBuilder::new()
+            .identification(identification())
+            .grid(mercator_grid(0))
+            .product(product(0, 0))
+            .packing(PackingStrategy::SimpleAuto { decimal_scale: 0 })
+            .values(&values)
+            .build()
+            .unwrap();
+
+        let file = GribFile::from_bytes(write_message([field])).unwrap();
+        let message = file.message(0).unwrap();
+        assert_eq!(message.grid_shape(), (3, 2));
+        assert_eq!(
+            message.projected_x_coordinates().unwrap(),
+            vec![0.0, 1_000.0, 2_000.0]
+        );
+        assert_eq!(
+            message.projected_y_coordinates().unwrap(),
+            vec![-0.0, -2_000.0]
+        );
+        match message.grid_definition() {
+            GridDefinition::Mercator(grid) => {
+                assert_eq!(grid.number_of_points, 6);
+                assert_eq!(grid.shape_of_earth, 6);
+                assert_eq!(grid.ni, 3);
+                assert_eq!(grid.nj, 2);
+                assert_eq!(grid.lat_first, -20_000_000);
+                assert_eq!(grid.lon_first, -100_000_000);
+                assert_eq!(grid.lat_d, 0);
+                assert_eq!(grid.lat_last, 20_000_000);
+                assert_eq!(grid.lon_last, -98_000_000);
+                assert_eq!(grid.di, 1_000_000);
+                assert_eq!(grid.dj, 2_000_000);
+            }
+            other => panic!("expected Mercator grid, got {other:?}"),
+        }
+        assert_eq!(message.read_flat_data_as_f64().unwrap(), values);
+    }
+
+    #[test]
     fn writes_lambert_conformal_grib2_field_readable_by_reader() {
         let values = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0];
         let field = Grib2FieldBuilder::new()
@@ -2565,6 +2780,49 @@ mod tests {
                 assert_eq!(grid.latin2, 25_000_000);
             }
             other => panic!("expected Lambert conformal grid, got {other:?}"),
+        }
+        assert_eq!(message.read_flat_data_as_f64().unwrap(), values);
+    }
+
+    #[test]
+    fn writes_albers_equal_area_grib2_field_readable_by_reader() {
+        let values = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0];
+        let field = Grib2FieldBuilder::new()
+            .identification(identification())
+            .grid(albers_grid(0))
+            .product(product(0, 0))
+            .packing(PackingStrategy::SimpleAuto { decimal_scale: 0 })
+            .values(&values)
+            .build()
+            .unwrap();
+
+        let file = GribFile::from_bytes(write_message([field])).unwrap();
+        let message = file.message(0).unwrap();
+        assert_eq!(message.grid_shape(), (3, 2));
+        assert_eq!(
+            message.projected_x_coordinates().unwrap(),
+            vec![0.0, 4_000.0, 8_000.0]
+        );
+        assert_eq!(
+            message.projected_y_coordinates().unwrap(),
+            vec![-0.0, -5_000.0]
+        );
+        match message.grid_definition() {
+            GridDefinition::AlbersEqualArea(grid) => {
+                assert_eq!(grid.number_of_points, 6);
+                assert_eq!(grid.shape_of_earth, 6);
+                assert_eq!(grid.nx, 3);
+                assert_eq!(grid.ny, 2);
+                assert_eq!(grid.lat_first, 23_000_000);
+                assert_eq!(grid.lon_first, 240_000_000);
+                assert_eq!(grid.lat_d, 25_000_000);
+                assert_eq!(grid.lon_v, 265_000_000);
+                assert_eq!(grid.dx, 4_000_000);
+                assert_eq!(grid.dy, 5_000_000);
+                assert_eq!(grid.latin1, 29_500_000);
+                assert_eq!(grid.latin2, 45_500_000);
+            }
+            other => panic!("expected Albers equal-area grid, got {other:?}"),
         }
         assert_eq!(message.read_flat_data_as_f64().unwrap(), values);
     }